Moved model_utils.py to utils/model_utils.py
